Refactor some code and fixed Vite build warnings
