+++ docs/index.html
@@ -15,7 +15,7 @@
                 window.location.hostname === "sygec.github.io" ||
                 window.location.hostname === "dicethrone-prod.sygec.workers.dev";
         </script>
-      <script type="module" crossorigin src="/assets/index-o2kDhndb.js"></script>
+      <script type="module" crossorigin src="/assets/index-DqWHwBMw.js"></script>
       <link rel="stylesheet" crossorigin href="/assets/index-D7vOqzRi.css">
     </head>
